requirements.txt
openai
dotenv
requests
pdfminer.six
python-docx
spacy